Update JSON form to support "binary" field
This updates the project.json format to support the "binary" field.  I
have added URLs for some (though not all) of the projects.

This also updates the downloader to support jar files.  And, updates the
RevAPI driver to work with them.  At this stage, the results are not
accurate (for reasons unknown).