Order versions in projects.json.  Updated RevAPI
This orders the versions in projects.json so they are now in
chronological order.  This also updates the RevAPI driver so that it
only reports BREAKING or POTENTIALLY_BREAKING changes.

There remains some issues however.  In particular, running with the same
library as the old and new jar still produces some delta.  This is
because some delta's are not specific to the two versions, but seem to
be errors which are always reported for a given jar file.  They maybe
related to unknown dependencies, perhaps.